+++ lib/membrane/core/child/pads_specs.ex
@@ -174,8 +174,8 @@ defmodule Membrane.Core.Child.PadsSpecs do
       |> Enum.filter(fn {_pad, info} -> info[:flow_control] in [:auto, :push] end)
       |> Enum.split_with(fn {_pad, info} -> info.flow_control == :auto end)
 
-    if auto_pads != [] and push_pads != [] do
-      IO.warn("""
+    if should_warn_on_mixing_push_and_auto_pads?(pads, auto_pads, push_pads, used_module) do
+      Membrane.Logger.warning("""
       #{inspect(env.module)} defines pads with `flow_control: :auto` and pads with `flow_control: :push` \
       at the same time. Please, consider if this what you want to do - flow control of these pads won't be \
       integrated. Setting `flow_control` to `:auto` in the places where it has been set to `:push` might be \
@@ -220,6 +220,26 @@ defmodule Membrane.Core.Child.PadsSpecs do
 
     :ok
   end
+
+  defp should_warn_on_mixing_push_and_auto_pads?(all_pads, auto_pads, push_pads, used_module) do
+    auto_pads != [] and push_pads != [] and
+      not endpoint_with_surely_valid_flow_control(used_module, all_pads)
+  end
+
+  # returns true if e.g. Endpoint has all input pads in auto and all output pads in push
+  defp endpoint_with_surely_valid_flow_control(Membrane.Endpoint, pads) do
+    pads
+    |> Enum.group_by(
+      fn {_pad_name, pad_info} -> pad_info.direction end,
+      fn {_pad_name, pad_info} -> pad_info end
+    )
+    |> Enum.all?(fn {_directiom, pads_info} ->
+      MapSet.new(pads_info, & &1.flow_control)
+      |> MapSet.size() == 1
+    end)
+  end
+
+  defp endpoint_with_surely_valid_flow_control(_used_module, _pads), do: false
 
   @spec parse_pad_specs!(
           specs :: Pad.spec(),