+++ lib/membrane/core/telemetry.ex
@@ -126,11 +126,16 @@ defmodule Membrane.Core.Telemetry do
     end
   end
 
+  @doc false
+  @spec identity(term()) :: term()
+  def identity(term), do: term
+
   defmacrop report_datapoint(datapoint_name, do: lazy_block) do
     unless Macro.quoted_literal?(datapoint_name), do: raise("Datapoint type must be a literal")
 
+    # @legacy? |> indentity() is a hack to mute compilation warnings
     cond do
-      @legacy? ->
+      @legacy? |> identity() ->
         do_legacy_telemetry(datapoint_name, lazy_block)
 
       datapoint_gathered?(datapoint_name) ->
@@ -171,15 +176,16 @@ defmodule Membrane.Core.Telemetry do
           Telemetry.callback_context()
         ) :: CallbackHandler.callback_return() | no_return()
   def track_callback_handler(f, callback, args, state, context) do
+    component_type = state_module_to_atom(state.__struct__)
+
     meta =
       callback_meta(
         state,
         callback,
         args,
-        context
+        context,
+        component_type
       )
-
-    component_type = state_module_to_atom(state.__struct__)
 
     if handler_reported?(component_type, callback) do
       :telemetry.span([:membrane, component_type, callback], meta, fn ->
@@ -193,13 +199,13 @@ defmodule Membrane.Core.Telemetry do
     end
   end
 
-  defp callback_meta(state, callback, args, context) do
+  defp callback_meta(state, callback, args, context, component_type) do
     %{
       callback: callback,
       callback_args: args,
       callback_context: context,
       component_path: ComponentPath.get(),
-      component_type: state.module,
+      component_type: component_type,
       monotonic_time: System.monotonic_time(),
       state_before_callback: state.internal_state,
       state_after_callback: nil