+++ lib/membrane/pipeline/callback_context.ex
@@ -11,8 +11,14 @@ defmodule Membrane.Pipeline.CallbackContext do
   Field `:start_of_stream_received?` is present only in
   `c:Membrane.Pipeline.handle_element_end_of_stream/4`.
 
-  Fields `:members`, `:crash_initiator` and `:crash_reason` are present only in
+  Field `:crash_initiator` is only present in `c:Membrane.Pipeline.handle_child_terminated/3`
+  and `c:Membrane.Pipeline.handle_crash_group_down/3`.
+
+  Fields `:members` and `:crash_reason` are present only in
   `c:Membrane.Pipeline.handle_crash_group_down/3`.
+
+  Fields `:exit_reason` and `:group_name` are present only in
+  `c:Membrane.Pipeline.handle_child_terminated/3`.
   """
   @type t :: %{
           :children => %{Membrane.Child.name() => Membrane.ChildEntry.t()},
@@ -24,8 +30,10 @@ defmodule Membrane.Pipeline.CallbackContext do
           :utility_supervisor => Membrane.UtilitySupervisor.t(),
           optional(:from) => [GenServer.from()],
           optional(:members) => [Membrane.Child.name()],
-          optional(:crash_initiator) => Membrane.Child.name(),
+          optional(:crash_initiator) => Membrane.Child.name() | nil,
           optional(:crash_reason) => :normal | :shutdown | {:shutdown, term()} | term(),
-          optional(:start_of_stream_received?) => boolean()
+          optional(:start_of_stream_received?) => boolean(),
+          optional(:exit_reason) => :normal | :shutdown | {:shutdown, term()} | term(),
+          optional(:group_name) => Membrane.Child.group() | nil
         }
 end