+++ scripts/elixir/update_packages_list.exs
@@ -2,137 +2,13 @@ Mix.install([{:req, "~> 0.4.0"}])
 
 require Logger
 
-# define packages structure
+{packages, _bindings} = Code.eval_file("packages.exs", __DIR__)
+
 packages =
-  [
-    {:md, "### General"},
-    "membrane_sdk",
-    "membrane_core",
-    "membrane_rtc_engine",
-    "kino_membrane",
-    "docker_membrane",
-    "membrane_demo",
-    "membrane_tutorials",
-    {:md, "### Plugins"},
-    {:md, "#### General purpose"},
-    "membrane_file_plugin",
-    "membrane_hackney_plugin",
-    "membrane_scissors_plugin",
-    "membrane_tee_plugin",
-    "membrane_funnel_plugin",
-    "membrane_realtimer_plugin",
-    "membrane_stream_plugin",
-    "membrane_fake_plugin",
-    "membrane_pcap_plugin",
-    "kim-company/membrane_live_framerate_converter_plugin",
-    "membrane_template_plugin",
-    {:md, "#### Streaming protocols"},
-    "membrane_webrtc_plugin",
-    "membrane_rtmp_plugin",
-    "membrane_http_adaptive_stream_plugin",
-    "membrane_ice_plugin",
-    "membrane_udp_plugin",
-    "membrane_tcp_plugin",
-    "membrane_rtp_plugin",
-    "membrane_rtp_h264_plugin",
-    "membrane_rtp_vp8_plugin",
-    "membrane_rtp_vp9_plugin",
-    "membrane_rtp_mpegaudio_plugin",
-    "membrane_rtp_opus_plugin",
-    "membrane_rtp_g711_plugin",
-    "mickel8/membrane_quic_plugin",
-    "kim-company/membrane_mpeg_ts_plugin",
-    "kim-company/membrane_hls_plugin",
-    {:md, "#### Containers"},
-    "membrane_mp4_plugin",
-    "membrane_matroska_plugin",
-    "membrane_flv_plugin",
-    "membrane_ivf_plugin",
-    "membrane_ogg_plugin",
-    {:md, "#### Audio codecs"},
-    "membrane_aac_plugin",
-    "membrane_aac_fdk_plugin",
-    "membrane_flac_plugin",
-    "membrane_mp3_lame_plugin",
-    "membrane_mp3_mad_plugin",
-    "membrane_opus_plugin",
-    "membrane_wav_plugin",
-    "membrane_g711_plugin",
-    {:md, "#### Video codecs"},
-    "membrane_h26x_plugin",
-    "membrane_h264_ffmpeg_plugin",
-    "membrane_vpx_plugin",
-    "membrane_abr_transcoder_plugin",
-    "gBillal/membrane_h265_ffmpeg_plugin",
-    "binarynoggin/elixir-turbojpeg",
-    "kim-company/membrane_subtitle_mixer_plugin",
-    {:md, "#### Raw audio & video"},
-    "membrane_generator_plugin",
-    {:md, "**Raw audio**"},
-    "membrane_raw_audio_parser_plugin",
-    "membrane_portaudio_plugin",
-    "membrane_audio_mix_plugin",
-    "membrane_audio_filler_plugin",
-    "membrane_ffmpeg_swresample_plugin",
-    "membrane_audiometer_plugin",
-    {:md, "**Raw video**"},
-    "membrane_raw_video_parser_plugin",
-    "membrane_video_merger_plugin",
-    "membrane_live_compositor_plugin",
-    "membrane_camera_capture_plugin",
-    "membrane_rpicam_plugin",
-    "membrane_framerate_converter_plugin",
-    "membrane_sdl_plugin",
-    "membrane_overlay_plugin",
-    "membrane_ffmpeg_swscale_plugin",
-    "membrane_ffmpeg_video_filter_plugin",
-    "kim-company/membrane_video_mixer_plugin",
-    {:md, "#### External APIs"},
-    "membrane_aws_plugin",
-    "membrane_agora_plugin",
-    "membrane_element_gcloud_speech_to_text",
-    "membrane_element_ibm_speech_to_text",
-    "YuzuTen/membrane_s3_plugin",
-    "lawik/membrane_transcription",
-    {:md, "### Formats"},
-    "membrane_rtp_format",
-    "membrane_cmaf_format",
-    "membrane_matroska_format",
-    "membrane_mp4_format",
-    "membrane_raw_audio_format",
-    "membrane_raw_video_format",
-    "membrane_aac_format",
-    "membrane_opus_format",
-    "membrane_flac_format",
-    "membrane_mpegaudio_format",
-    "membrane_h264_format",
-    "membrane_vp8_format",
-    "membrane_vp9_format",
-    "membrane_g711_format",
-    {:md, "### Standalone media libs"},
-    "elixir-webrtc/ex_webrtc",
-    "ex_sdp",
-    "ex_libnice",
-    "ex_libsrtp",
-    "ex_m3u8",
-    "membrane_rtsp",
-    "membrane_ffmpeg_generator",
-    {:md, "### Utils"},
-    "unifex",
-    "bundlex",
-    "beamchmark",
-    "bunch",
-    "bunch_native",
-    "shmex",
-    "membrane_timestamp_queue",
-    "membrane_common_c",
-    "membrane_telemetry_metrics",
-    "membrane_opentelemetry",
-    "membrane_precompiled_dependency_provider"
-  ]
+  packages
   |> Enum.map(fn
-    {:md, markdown} ->
-      %{type: :markdown, content: markdown}
+    {type, markdown} ->
+      %{type: type, name: markdown}
 
     package when is_binary(package) ->
       case String.split(package, "/", parts: 2) do
@@ -147,16 +23,25 @@ gh_req_timeout = 500
 # for debugging, allows mocking requests for particular repos
 gh_req_mock = false
 
+# gh token for larger request rate
+gh_auth_headers =
+  case System.get_env("GITHUB_TOKEN") do
+    nil -> []
+    token -> [Authorization: "Bearer #{token}"]
+  end
+
 # fetch repos from the known organizations
 repos =
   ["membraneframework", "membraneframework-labs", "fishjam-dev"]
   |> Enum.flat_map(fn org ->
     Stream.from_index()
     |> Stream.map(fn page ->
-      Process.sleep(gh_req_timeout)
+      if gh_auth_headers == [], do: Process.sleep(gh_req_timeout)
       url = "https://api.github.com/orgs/#{org}/repos?per_page=100&page=#{page}"
       Logger.debug("Fetching #{url}")
-      resp = Req.get!(url, decode_json: [keys: :atoms]).body
+
+      resp =
+        Req.get!(url, headers: gh_auth_headers, decode_json: [keys: :atoms]).body
 
       unless is_list(resp) do
         raise "Received invalid response: #{inspect(resp)}"
@@ -184,8 +69,8 @@ packages_blacklist = [
   ".github",
   "membraneframework.github.io",
   "membrane_rtc_engine_timescaledb",
-  "membrane_g711_ffmpeg_plugin",
-  "github_actions_test"
+  "github_actions_test",
+  "membrane_ice_plugin"
 ]
 
 lacking_repos =
@@ -215,10 +100,11 @@ packages =
             %{owner: %{login: :mock}, html_url: :mock, description: :mock}
 
           owner != nil ->
-            Process.sleep(gh_req_timeout)
+            if gh_auth_headers == [], do: Process.sleep(gh_req_timeout)
             url = "https://api.github.com/repos/#{owner}/#{name}"
             Logger.debug("Fetching #{url}")
-            Req.get!(url, decode_json: [keys: :atoms]).body
+
+            Req.get!(url, headers: gh_auth_headers, decode_json: [keys: :atoms]).body
 
           Map.has_key?(repos, name) ->
             Map.fetch!(repos, name)
@@ -228,14 +114,35 @@ packages =
         end
 
       hex = Req.get!("https://hex.pm/api/packages/#{name}", decode_json: [keys: :atoms])
-      is_hex_present = hex.status == 200
+
+      hex_badge =
+        if hex.status == 200 and hex.body.url != nil do
+          "[![Hex.pm](https://img.shields.io/hexpm/v/#{name}.svg)](#{hex.body.url})"
+        end
+
+      hexdocs_badge =
+        if hex.status == 200 and hex.body.docs_html_url != nil do
+          "[![Docs](https://img.shields.io/badge/api-docs-yellow.svg?style=flat)](#{hex.body.docs_html_url})"
+        end
+
+      github_badge =
+        "[![GitHub](https://img.shields.io/badge/github-code-white.svg?logo=github)](#{repo.html_url})"
+
+      owner_prefix =
+        case repo.owner.login do
+          "membraneframework-labs" -> "[Labs] "
+          "membraneframework" -> ""
+          owner -> "[Maintainer: [#{owner}](https://github.com/#{owner})] "
+        end
 
       Map.merge(package, %{
         owner: repo.owner.login,
         url: repo.html_url,
         description: repo.description,
-        hex_url: if(is_hex_present, do: hex.body.url),
-        hexdocs_url: if(is_hex_present, do: hex.body.docs_html_url)
+        owner_prefix: owner_prefix,
+        hex_badge: hex_badge,
+        hexdocs_badge: hexdocs_badge,
+        github_badge: github_badge
       })
 
     other ->
@@ -250,46 +157,36 @@ header = """
 | --- | --- | --- |
 """
 
+generated_code_comment =
+  "<!-- Generated code, do not edit. See `scripts/elixir/update_packages_list.exs`. -->"
+
 packages_md =
   packages
-  |> Enum.map_reduce(%{is_header_present: false}, fn
-    %{type: :markdown, content: content}, acc ->
-      {"\n#{content}", %{acc | is_header_present: false}}
+  |> Enum.map_reduce(
+    %{is_header_present: false},
+    fn
+      %{type: :section, name: name}, acc ->
+        {"\n### " <> name, %{acc | is_header_present: false}}
 
-    %{type: :package} = package, acc ->
-      prefix =
-        case package.owner do
-          "membraneframework-labs" -> "[Labs] "
-          "membraneframework" -> ""
-          _other -> "[Maintainer: [#{package.owner}](https://github.com/#{package.owner})] "
-        end
+      %{type: :subsection, name: name}, acc ->
+        {"\n#### " <> name, %{acc | is_header_present: false}}
 
-      hex_badge =
-        if package.hex_url,
-          do:
-            "[![Hex.pm](https://img.shields.io/hexpm/v/#{package.name}.svg)](#{package.hex_url})"
+      %{type: :package} = package, acc ->
+        package_info = """
+        #{if acc.is_header_present, do: "", else: header}\
+        | [#{package.name}](#{package.url}) | #{package.owner_prefix}#{package.description} | #{package.hex_badge} #{package.hexdocs_badge} |\
+        """
 
-      hexdocs_badge =
-        if package.hexdocs_url,
-          do:
-            "[![Docs](https://img.shields.io/badge/api-docs-yellow.svg?style=flat)](#{package.hexdocs_url})"
-
-      url = "[#{package.name}](#{package.url})"
-
-      result = """
-      #{if acc.is_header_present, do: "", else: header}\
-      | #{url} | #{prefix}#{package.description} | #{hex_badge} #{hexdocs_badge} |\
-      """
-
-      {result, %{acc | is_header_present: true}}
-  end)
+        {package_info, %{acc | is_header_present: true}}
+    end
+  )
   |> elem(0)
   |> Enum.join("\n")
 
 packages_md =
   """
   <!-- packages-list-start -->
-  <!-- Generated code, do not edit. See `scripts/elixir/update_packages_list.exs`. -->
+  #{generated_code_comment}
 
   #{packages_md}
 
@@ -306,5 +203,60 @@ File.read!(readme_path)
   packages_md
 )
 |> then(&File.write!(readme_path, &1))
+
+# update packages in docs
+packages_docs_path = "guides/packages"
+
+File.rm_rf!(packages_docs_path)
+File.mkdir_p(packages_docs_path)
+
+packages
+|> Enum.reduce(
+  %{file_path: nil, file_number: 0, section: nil, files: []},
+  fn
+    %{type: :package} = package, acc ->
+      package_info = """
+      ## #{package.name}
+      #{package.owner_prefix}#{package.description} 
+
+      #{package.hex_badge} #{package.hexdocs_badge} #{package.github_badge}
+       
+      """
+
+      files =
+        List.update_at(acc.files, 0, fn {file_path, file_content} ->
+          {file_path, file_content <> package_info}
+        end)
+
+      %{acc | files: files}
+
+    %{type: type, name: name}, acc ->
+      # So that the files have correct order
+      prefix = "#{acc.file_number}_" |> String.pad_leading(3, "0")
+
+      {filename, section} =
+        case type do
+          :section -> {name, name}
+          :subsection -> {"#{acc.section} | #{name}", acc.section}
+        end
+
+      file_path =
+        Path.join(packages_docs_path, prefix <> filename <> ".md") |> String.replace(" ", "_")
+
+      files = [{file_path, ""} | acc.files]
+
+      %{
+        acc
+        | file_path: file_path,
+          section: section,
+          file_number: acc.file_number + 1,
+          files: files
+      }
+  end
+)
+|> Map.get(:files)
+|> Enum.each(fn {file_path, file_content} ->
+  if file_content != "", do: File.write!(file_path, "#{generated_code_comment}\n#{file_content}")
+end)
 
 IO.puts("Packages updated successfully.")